add confirmation text hook
