+++ src/views/GovernanceTaskList/index.tsx
@@ -67,7 +67,7 @@ const GovernanceTaskList = () => {
   const archiveIntake = () => {
     const redirect = () => {
       history.push('/', {
-        confirmationText: t('taskList:withdraw_modal:text', {
+        confirmationText: t('taskList:withdraw_modal.confirmationText', {
           context: systemIntake.requestName ? 'name' : 'noName',
           requestName: systemIntake.requestName
         })
